# docs/openapi.yaml
openapi: 3.0.0
info:
  title: SkillYards API
  description: API for Student Management, Fee Ledgers, and Testing Sessions.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development Server
paths:
...
  /api/students/{id}/plan:
    get:
      tags: [Accounting]
      summary: Get active fee plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWrapper'
    post:
      tags: [Accounting]
      summary: Assign a new fee plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlan'
      responses:
        '201':
          description: Plan assigned
...
components:
  schemas:
    ApiResponseWrapper:
      type: object
      properties:
        success: { type: boolean }
        data: { type: object }
        message: { type: string }
...
    CreatePlan:
      type: object
      required: [type, totalAmount]
      properties:
        type: { type: string, enum: [full, emi, custom, flexible] }
        totalAmount: { type: integer }
        installments:
          type: array
          items:
            type: object
            properties:
              amount: { type: integer }
              dueDate: { type: string, format: date-time }
...